Move Page class to separate file
